Prioritize hosts by current usage rather than just lifetime
We kill our hosts regularly, and so we would like to schedule drivers so
that they heuristically get the longest remaining lifetime. For this
reason, we currently sort our hosts purely in terms of their remaining
host lifetime.

In practice, this is demonstrably wasteful. Consider the case where we
have hosts A and B, where A has a single pod running on it and we've
just booted up host B and so it has no pods running on it. An arriving
job will be scheduled onto host B, leading to poor usage. We would
prefer to schedule the arriving job onto A as well, and be able to kill
B.

This PR takes current usage into account. First, it attempts to fill up
old nodes, regardless of their usage. For young nodes (the ones where
suitability for scheduling drivers is immaterial), we instead schedule
first based on memory (prefer the hosts with the least remaining memory)
and then CPU (prefer the hosts with the least remaining CPU).

The goal is that instead of having an even spread of partially used
hosts, we'd be able to partition into totally used hosts and mostly
unused hosts.